semantics

+++ zeta0/main.c
@@ -8,7 +8,7 @@ int main(int argc, char **argv) {
 	int n = atoi(argv[1]);
 	double sum = 0;
 	for (int i = 1; i <= n; i++) {
-		sum += (double) 1/(i*i);
+		sum += 1.0/(i*i);
 	}
 
 	double num_pi = sqrt(6*sum);
